Good Nestjs

+++ src/api/boards/board.service.ts
@@ -4,6 +4,7 @@ import { CreateBoardDto } from './dto/create-board.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { BoardRepository } from './board.repository';
 import { Board } from './board.entity';
+import { Auth } from '../auth/auth.entity';
 
 @Injectable()
 export class BoardService {
@@ -12,8 +13,14 @@ export class BoardService {
         private boardRepository: BoardRepository,
     ) {}
     getBoardById(id: number): Promise<Board> {
-        console.log(id);
         const result = this.boardRepository.findOne(id);
+        if (!result) {
+            throw new NotFoundException(`Can't find Board with id ${id}`);
+        }
+        return result;
+    }
+    getBoardByIdUpdate(id: number, auth: Auth): Promise<Board> {
+        const result = this.boardRepository.findOne({ id, auth });
         if (!result) {
             throw new NotFoundException(`Can't find Board with id ${id}`);
         }
@@ -26,21 +33,35 @@ export class BoardService {
         }
         return result;
     }
-    createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
-        return this.boardRepository.createBoard(createBoardDto);
+    createBoard(createBoardDto: CreateBoardDto, auth: Auth): Promise<Board> {
+        return this.boardRepository.createBoard(createBoardDto, auth);
     }
-    async deleteBoard(id: number): Promise<string> {
-        const result = await this.boardRepository.delete(id);
+    async deleteBoard(id: number, auth: Auth): Promise<string> {
+        const result = await this.boardRepository.delete({ id, auth });
         if (result.affected == 0) {
             throw new NotFoundException(`Can't find Board with id ${id}`);
         }
         return `Delete Board with id ${id}`;
     }
-    async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-        const board = await this.getBoardById(id);
+    async updateBoardStatus(
+        id: number,
+        status: BoardStatus,
+        auth: Auth,
+    ): Promise<Board> {
+        const board = await this.getBoardByIdUpdate(id, auth);
+        if (!board) {
+            throw new NotFoundException(`Can't find Board with id ${id}`);
+        }
         board.status = status;
         await this.boardRepository.save(board);
         return board;
+    }
+    async getMyBoard(auth: Auth): Promise<Board[]> {
+        const result = await this.boardRepository.find({ auth: auth });
+        if (!result || result[0] === undefined) {
+            throw new NotFoundException(`Can't find My Board all`);
+        }
+        return result;
     }
     // getBoardById(id: string): Board {
     //     const result = this.boards.find((board) => board.id === id);
